Consolidated store on top-level
Having the store modules with the components that use them makes
a lot of sense on paper but in practice it makes finding a module
often more difficult.

This consolidates all store modules and plugins in the src directory
for a better overview.
