+++ src/modules/drawing/components/DrawingToolboxButton.vue
@@ -14,9 +14,9 @@
 </template>
 
 <script>
-import { drawingModes } from '@/modules/store/modules/drawing.store'
+import { drawingModes } from '@/store/modules/drawing.store'
 import ButtonWithIcon from '@/utils/ButtonWithIcon.vue'
-import { UIModes } from '@/modules/store/modules/ui.store'
+import { UIModes } from '@/store/modules/ui.store'
 
 export default {
     components: { ButtonWithIcon },
