--- src/modules/i18n/store/index.js
@@ -1,3 +1,0 @@
-import i18nStore from './i18n.store'
-
-export default i18nStore
