+++ src/modules/map/components/LocationPopup.vue
@@ -105,7 +105,7 @@ import { mapState, mapActions } from 'vuex'
 import { registerWhat3WordsLocation } from '@/api/what3words.api'
 import { requestHeight } from '@/api/height.api'
 import { generateQrCode } from '@/api/qrcode.api'
-import { ClickType } from '@/modules/map/store/map.store'
+import { ClickType } from '@/store/modules/map.store'
 import OpenLayersPopover from '@/modules/map/components/openlayers/OpenLayersPopover.vue'
 import LocationPopupCopyInput from '@/modules/map/components/LocationPopupCopyInput.vue'
 import LocationPopupCopySlot from '@/modules/map/components/LocationPopupCopySlot.vue'
