+++ src/modules/map/components/openlayers/OpenLayersMap.vue
@@ -88,8 +88,8 @@ import { Feature } from '@/api/features.api'
 import LayerTypes from '@/api/layers/LayerTypes.enum'
 import HighlightedFeatureList from '@/modules/infobox/components/HighlightedFeatureList.vue'
 import OpenLayersPopover from '@/modules/map/components/openlayers/OpenLayersPopover.vue'
-import { ClickInfo, ClickType } from '@/modules/map/store/map.store'
-import { CrossHairs } from '@/modules/store/modules/position.store'
+import { ClickInfo, ClickType } from '@/store/modules/map.store'
+import { CrossHairs } from '@/store/modules/position.store'
 import ButtonWithIcon from '@/utils/ButtonWithIcon.vue'
 import log from '@/utils/logging'
