--- src/modules/map/store/index.js
@@ -1,3 +1,0 @@
-import mapStore from './map.store'
-
-export default mapStore
