--- src/modules/menu/store/index.js
@@ -1,3 +1,0 @@
-import search from './search.store'
-
-export default search
