+++ src/store/index.js
@@ -1,17 +1,16 @@
 import { createStore } from 'vuex'
 
 import app from './modules/app.store'
-import layers from './modules/layers.store'
-import geolocation from './modules/geolocation.store'
-import position from './modules/position.store'
-import topics from './modules/topics.store'
-import ui from './modules/ui.store'
 import drawing from './modules/drawing.store'
 import feature from './modules/feature.store'
-
-import map from '@/modules/map/store'
-import search from '@/modules/menu/store'
-import i18n from '@/modules/i18n/store'
+import geolocation from './modules/geolocation.store'
+import i18n from './modules/i18n.store'
+import layers from './modules/layers.store'
+import map from './modules/map.store'
+import position from './modules/position.store'
+import search from './modules/search.store'
+import topics from './modules/topics.store'
+import ui from './modules/ui.store'
 
 import loadLayersConfigOnLangChange from './plugins/load-layersconfig-on-lang-change'
 import redoSearchOnLangChange from './plugins/redo-search-on-lang-change.plugin'
@@ -19,10 +18,10 @@ import menuSearchBarInteractionManagementPlugin from './plugins/menu-search-inte
 import clickOnMapManagementPlugin from './plugins/click-on-map-management.plugin'
 import appReadinessPlugin from './plugins/app-readiness.plugin'
 import geolocationManagementPlugin from './plugins/geolocation-management.plugin'
-import topicChangeManagementPlugin from '@/modules/store/plugins/topic-change-management.plugin'
-import loadingBarManagementPlugin from '@/modules/store/plugins/loading-bar-management.plugin'
-import drawingLayerManagementPlugin from '@/modules/store/plugins/drawing-layer-management.plugin'
-import screenSizeManagementPlugin from '@/modules/store/plugins/screen-size-management.plugin'
+import topicChangeManagementPlugin from './plugins/topic-change-management.plugin'
+import loadingBarManagementPlugin from './plugins/loading-bar-management.plugin'
+import drawingLayerManagementPlugin from './plugins/drawing-layer-management.plugin'
+import screenSizeManagementPlugin from './plugins/screen-size-management.plugin'
 
 export default createStore({
     strict: true,
