+++ src/store/modules/drawing.store.js
