+++ src/store/plugins/drawing-layer-management.plugin.js
