+++ src/store/plugins/geolocation-management.plugin.js
