+++ src/store/plugins/loading-bar-management.plugin.js
