+++ src/store/plugins/menu-search-interaction.plugin.js
