+++ src/store/plugins/screen-size-management.plugin.js
@@ -1,4 +1,4 @@
-import { UIModes } from '@/modules/store/modules/ui.store'
+import { UIModes } from '@/store/modules/ui.store'
 import { BREAKPOINT_PHONE_WIDTH, BREAKPOINT_PHONE_HEIGHT, BREAKPOINT_TABLET } from '@/config'
 
 /** @param store */
